+++ projects/_03_independent_cqc/unittest_data/polars_ind_cqc_test_file_schemas.py
@@ -1836,3 +1836,56 @@ class EstimateFilledPostsByJobRoleCleanSchemas:
         IndCQC.provider_id: CatColType.ProviderCatType,
         IndCQC.brand_id: CatColType.BrandCatType,
     }
+
+
+@dataclass
+class CombineASCWDSAndPIRSchemas:
+    blend_pir_and_ascwds_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.care_home: pl.String,
+        IndCQC.pir_filled_posts_model: pl.Float32,
+        IndCQC.ascwds_filled_posts_dedup_clean: pl.Float32,
+    }
+
+    expected_create_repeated_ascwds_clean_column_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.ascwds_filled_posts_dedup_clean: pl.Float32,
+        IndCQC.ascwds_filled_posts_dedup_clean_repeated: pl.Float32,
+    }
+
+    expected_create_last_submission_columns_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.ascwds_filled_posts_dedup_clean: pl.Float32,
+        IndCQC.pir_filled_posts_model: pl.Float32,
+        IndCQC.last_ascwds_submission: pl.Date,
+        IndCQC.last_pir_submission: pl.Date,
+    }
+
+    expected_create_ascwds_pir_merged_column_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.last_ascwds_submission: pl.Date,
+        IndCQC.last_pir_submission: pl.Date,
+        IndCQC.ascwds_filled_posts_dedup_clean_repeated: pl.Float32,
+        IndCQC.pir_filled_posts_model: pl.Float32,
+        IndCQC.ascwds_filled_posts_dedup_clean: pl.Float32,
+        IndCQC.ascwds_pir_merged: pl.Float32,
+    }
+
+    include_pir_if_never_submitted_ascwds_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.cqc_location_import_date: pl.Date,
+        IndCQC.ascwds_pir_merged: pl.Float32,
+        IndCQC.pir_filled_posts_model: pl.Float32,
+    }
+
+    drop_temporary_columns_schema = {
+        IndCQC.location_id: pl.String,
+        IndCQC.last_ascwds_submission: pl.Date,
+        IndCQC.last_pir_submission: pl.Date,
+        IndCQC.ascwds_filled_posts_dedup_clean_repeated: pl.Float32,
+    }
+    expected_drop_temporary_columns = [IndCQC.location_id]