+++ projects/_03_independent_cqc/utils/imputation/imputation.py
@@ -101,19 +101,29 @@ def split_dataset_for_imputation(
             - imputation_lf: LazyFrame with rows meeting the criteria for imputation.
             - non_imputation_lf: LazyFrame with rows not meeting the criteria.
     """
+
     if care_home:
         care_home_filter_expr: pl.Expr = is_care_home()
     else:
         care_home_filter_expr: pl.Expr = is_not_care_home()
 
-    locs_with_values_expr = (
-        pl.col(column_with_null_values)
-        .count()
-        .over([IndCqc.location_id, IndCqc.care_home])
-        > 0
-    ) & (care_home_filter_expr)
+    groups_with_values = (
+        lf.filter(pl.col(column_with_null_values).is_not_null())
+        .filter(care_home_filter_expr)
+        .select([IndCqc.location_id, IndCqc.care_home])
+        .unique()
+    )
 
-    imputation_lf = lf.filter(locs_with_values_expr == True)
-    non_imputation_lf = lf.filter(locs_with_values_expr == False)
+    imputation_lf = lf.join(
+        groups_with_values,
+        on=[IndCqc.location_id, IndCqc.care_home],
+        how="semi",
+    )
+
+    non_imputation_lf = lf.join(
+        groups_with_values,
+        on=[IndCqc.location_id, IndCqc.care_home],
+        how="anti",
+    )
 
     return (imputation_lf, non_imputation_lf)